requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0

# Text-to-speech and audio processing
gtts>=2.3.0
pydub>=0.25.0

# Translation capabilities
deep-translator>=1.11.0

# AI text generation
transformers>=4.30.0
torch>=2.0.0

# Advanced audio processing (optional)
librosa>=0.10.0
soundfile>=0.12.0

# Speech recognition (optional)
SpeechRecognition>=3.10.0

# Core dependencies
Pillow>=10.0.0
numpy>=1.24.0
networkx>=3.1
